Create Alara (NaN values, histogram, scatter plot, horizantal line)
With this code, I initially traversed the columns in a dataframe named "train" to identify only those columns with the "float64" data type. Next, I saved the list of these columns in an empty list called "n_floats".
Next, I created a list called "non_floats" and stored the columns that do not have the data type "float64" in that list. Finally, I created a dataframe named "train_2" and removed the columns in the list "non_floats" from this dataframe. Thus, the "train_2" data frame contained only columns with the "float64" data type, and confusion was avoided.
This code will create a histogram of the '62' column in the "train" dataset using the "plt.hist()" function. The histogram will be plotted with the color of the bars set to red using the "color=['red']" parameter. Finally, the histogram will be displayed using the "plt.show()" function.
Then I wrote a code that includes Finding the number of columns containing NaN values, and if NaN values are still in the data or not. Then I created a histogram and scatter plot. I showed the histograms of the 62nd and 120th columns in the same graph because the positive correlations are in columns 62 and 120.